*** empty log message ***

+++ Despe/WEB-INF/src/fr/umlv/desperados/stylesheet/DatabaseStyleSheetManager.java
@@ -10,6 +10,7 @@ import java.util.Map;
 import java.io.File;
 
 import fr.umlv.desperados.database.DatabaseRequestor;
+import fr.umlv.desperados.stylesheet.UsedStylesheetException;
 
 /**
  * Provides an implementation of the StyleSheetManager interface, using an 
@@ -74,12 +75,16 @@ public class DatabaseStyleSheetManager implements StyleSheetManager {
 	 * @param styleSheetId
 	 * @roseuid 3FF869D00394
 	 */
-	public void removeStyleSheet(String styleSheetId, String path) {
+	public void removeStyleSheet(String styleSheetId, String path) throws UsedStylesheetException {
 		try {
-			requestor.doQuery(
-				"DELETE FROM FEUILLE_DE_STYLE WHERE NOM_FIC_FEU='" + styleSheetId+"'");
-			File f = new File(path+styleSheetId);
-			f.delete();
+			ResultSet rs = requestor.doQuery("select * from DOCUMENT where FEUILLE_STYLE_DOC='"+styleSheetId+"'");
+			if(!rs.first()) {
+				requestor.doQuery(
+					"DELETE FROM FEUILLE_DE_STYLE WHERE NOM_FIC_FEU='" + styleSheetId+"'");
+				File f = new File(path+styleSheetId);
+				f.delete();
+			}
+			else throw new UsedStylesheetException("La feuille de style que vous voulez supprimer est actuellement utilisée.");
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
